make sure a2s.php is loaded when rendering
It is possible that the render fucntion is called while
the parse function was not called. This happens for example
when an edit is cancelled. Since a2s.php is loaded only
in handle(), all edit cancelled result in a2s.php not beeing
loaded, but beeing needed. => PHP error => blank dokuwiki
page. :(